fixed number of warnings
git-svn-id: https://open-ms.svn.sourceforge.net/svnroot/open-ms/OpenMS@4978 6adb6e08-d915-0410-941f-83917bcadc18

+++ source/APPLICATIONS/TOPP/RTPredict.C
@@ -333,11 +333,11 @@ class TOPPRTPredict
 			}
 			if (svm.getIntParameter(KERNEL_TYPE) == OLIGO)
 			{
-				number_of_peptides = modified_peptides.size();
+				number_of_peptides = (Int)modified_peptides.size();
 			}
 			else
 			{
-				number_of_peptides = peptides.size();
+				number_of_peptides = (Int)peptides.size();
 			}			
 										
 			vector<DoubleReal> rts;
